Fix "datatime" typo and body, and add two more snippets.
Assuming https://github.com/honza/vim-snippets was used as the
inspiration for this repo, I change 'datatime' to 'datetime' and removed
the 'T' from its body, matching the definition in vim-snippets.

I also added an new date snippet for the format common in the United
States, as well as a time snippet with seconds.