add/improve C snippets for IO and memory
Make these snippets more useful.

Also, I noticed that the snippets have been already using `"",` instead of `\n` before me, then I realized that `\n` is counter-intuitive and obfuscates the understanding of the snippets source, as opposed to `"",` where it's obvious that it's an empty line. So, I don't think it's a "dead space", but rather a necessary "separator" to better understand the snippets source and someone's final code.  
However, if you disagree, I would sadly :pensive: turn `"",` back to `\n` .