Update perl snippets (#460)
* include perl in package.json to trigger perl snippets
* fixed/updated some perl snippets
* Escape $$ and  $! occurrences
* With native neovim snippets $0 is buggy

Ok, apologies for multiple commits, but $0 placement is very buggy in neovim native snippets usage. The placeholder text for ${0:placeholder text} is not replaced when you actually start typing at that location. Instead of waiting for a fix in neovim, native-snippets.nvim or whatever tool, I'd rather just work around this in the snippets I'm writing.

* fix join command