feat(python): change order of navigation in comprehension snippets
This change is motivated by getting better completions when using a
language server. When starting from the `iterable` part of the snippet,
the language server can infer the type of the values contained in the
iterable, so that when we get to the part of writing the `value` it can
offer autocompletion and suggestions for fields and methods of `value`.