Add elixir snippets to define functions with args
The current snippets for defining functions don't have any tabstops in
between the opening and closing parens () of a function definition, so
if you want to define a function that has arguments, you won't be able
to tab your way into writing them.

The new snippets append an `a` to their respective "inspiration"
snippets, to indicate that they'll have a tabstop for easy [a]rguments.